+++ kitty/cocoa_window.m
@@ -267,6 +267,8 @@ PENDING(hide_macos_app, HIDE)
 PENDING(hide_macos_other_apps, HIDE_OTHERS)
 PENDING(minimize_macos_window, MINIMIZE)
 PENDING(quit, QUIT)
+PENDING(paste_from_clipboard, PASTE_FROM_CLIPBOARD)
+PENDING(copy_or_noop, COPY_OR_NOOP)
 
 - (BOOL)validateMenuItem:(NSMenuItem *)item {
     if (item.action == @selector(toggle_macos_secure_keyboard_entry:)) {
@@ -290,6 +292,21 @@ PENDING(quit, QUIT)
         item.action == @selector(detach_tab:))
     {
         if (![NSApp keyWindow]) return NO;
+    } else if (item.action == @selector(paste_from_clipboard:)) {
+        if (![NSApp keyWindow]) return NO;
+        NSPasteboard *pb = [NSPasteboard generalPasteboard];
+        if (![pb stringForType:NSPasteboardTypeString]) return NO;
+    } else if (item.action == @selector(copy_or_noop:)) {
+        if (![NSApp keyWindow]) return NO;
+        OSWindow *osw = current_os_window();
+        if (osw && osw->num_tabs > osw->active_tab) {
+            Tab *tab = osw->tabs + osw->active_tab;
+            if (tab->num_windows > tab->active_window) {
+                Screen *screen = tab->windows[tab->active_window].render_data.screen;
+                if (screen && screen_has_selection(screen)) return YES;
+            }
+        }
+        return NO;
     }
     return YES;
 }
@@ -323,6 +340,7 @@ typedef struct {
     GlobalShortcut toggle_macos_secure_keyboard_entry, toggle_fullscreen, open_kitty_website;
     GlobalShortcut hide_macos_app, hide_macos_other_apps, minimize_macos_window, quit, search_scrollback;
     GlobalShortcut macos_cycle_through_os_windows, macos_cycle_through_os_windows_backwards;
+    GlobalShortcut paste_from_clipboard, copy_or_noop;
 } GlobalShortcuts;
 static GlobalShortcuts global_shortcuts;
 
@@ -342,6 +360,7 @@ cocoa_set_global_shortcut(PyObject *self UNUSED, PyObject *args) {
     else Q(open_kitty_website); else Q(hide_macos_app); else Q(hide_macos_other_apps);
     else Q(minimize_macos_window); else Q(quit); else Q(search_scrollback);
     else Q(macos_cycle_through_os_windows); else Q(macos_cycle_through_os_windows_backwards);
+    else Q(paste_from_clipboard); else Q(copy_or_noop);
 #undef Q
     if (gs == NULL) { PyErr_SetString(PyExc_KeyError, "Unknown shortcut name"); return NULL; }
     int cocoa_mods;
@@ -797,6 +816,9 @@ cocoa_create_global_menu(void) {
     MENU_ITEM(editMenu, @"Clear Screen", clear_screen);
     MENU_ITEM(editMenu, @"Clear Last Command", clear_last_command);
     MENU_ITEM(editMenu, @"Find", search_scrollback);
+    [editMenu addItem:[NSMenuItem separatorItem]];
+    MENU_ITEM(editMenu, @"Copy", copy_or_noop);
+    MENU_ITEM(editMenu, @"Paste", paste_from_clipboard);
     [editMenu release];
 
     NSMenuItem* windowMenuItem =